+++ src/export/channel.rs
@@ -2,40 +2,6 @@ use crate::{scheduler::Runtime, utils, Event, Message};
 use std::sync::Arc;
 use tracing::{debug, error, info};
 
-fn store_if(runtime: &Arc<Runtime>, ack: bool, chan_id: &str, pattern: &str, message: &Message) {
-    if ack && !chan_id.is_empty() && message.retry_times == 0 {
-        println!("store: {message:?}");
-        let msg = message.into(chan_id, pattern);
-        runtime
-            .cache()
-            .store()
-            .base()
-            .messages()
-            .create(&msg)
-            .unwrap_or_else(|err| {
-                error!("channel.store_if_emit_id: {}", err.to_string());
-                eprintln!("channel.store_if_emit_id: {}", err);
-                false
-            });
-    }
-}
-
-fn is_match(
-    glob: &(
-        globset::GlobMatcher,
-        globset::GlobMatcher,
-        globset::GlobMatcher,
-        globset::GlobMatcher,
-    ),
-    e: &Event<Message>,
-) -> bool {
-    let (pat_type, pat_state, pat_tag, pat_key) = glob;
-    pat_type.is_match(&e.r#type)
-        && pat_state.is_match(e.state.as_ref())
-        && (pat_tag.is_match(&e.tag) || pat_tag.is_match(&e.model.tag))
-        && pat_key.is_match(&e.key)
-}
-
 #[derive(Debug, Clone)]
 pub struct ChannelOptions {
     pub id: String,
@@ -207,4 +173,38 @@ impl Channel {
     pub fn close(&self) {
         self.runtime.emitter().remove(&self.chan_id);
     }
+}
+
+fn store_if(runtime: &Arc<Runtime>, ack: bool, chan_id: &str, pattern: &str, message: &Message) {
+    if ack && !chan_id.is_empty() && message.retry_times == 0 {
+        println!("store: {message:?}");
+        let msg = message.into(chan_id, pattern);
+        runtime
+          .cache()
+          .store()
+          .base()
+          .messages()
+          .create(&msg)
+          .unwrap_or_else(|err| {
+              error!("channel.store_if_emit_id: {}", err.to_string());
+              eprintln!("channel.store_if_emit_id: {}", err);
+              false
+          });
+    }
+}
+
+fn is_match(
+    glob: &(
+        globset::GlobMatcher,
+        globset::GlobMatcher,
+        globset::GlobMatcher,
+        globset::GlobMatcher,
+    ),
+    e: &Event<Message>,
+) -> bool {
+    let (pat_type, pat_state, pat_tag, pat_key) = glob;
+    pat_type.is_match(&e.r#type)
+      && pat_state.is_match(e.state.as_ref())
+      && (pat_tag.is_match(&e.tag) || pat_tag.is_match(&e.model.tag))
+      && pat_key.is_match(&e.key)
 }
